fix errorneous local storage migration
only remove files when trying to copy stuff in a local recording
migration if and only if there's a source file that we're going to be
copying over. this prevents a situation where the user moves the local
folder in windows explorer and then tries to switch the folder in
squadov. previously, we'd just wipe all the files because the source
folders no longer exist.

fixes #1867