remove duplicate local disk space refresh calls
the primary app is created along with each new router view due to weird
things with how the Vue router is setup at the root level. thus the
alert as is was being created multiple times and therefore, each new
page a user navigates to will check the disk space AND create an
interval to continuously check that space. move the disk space check
into an interval at the app level and store the disk space remaining in
the Vuex store and have the component only check that value instead of
computing it.

also increase the amount of time between checks to reduce the amount of
log lines the service log will increase by

fixes #1904