made process running state object more reliable in the face of pid reuse
also added tests to ensure this behavior is correct. this is effectively
going back to the old functionality from bbf620d6 which was for #1392. I
think the primary difference is that the previous iteration did a pretty
inefficient search for processes based on a binary search whereas now we
have a hashmap so I don't see the CPU usage that was mentioned in the
old issue.

fixes #1716