better handling of pre-existing WoW combat log files
there technically should be little difference in handling a legacy wow
combat log file vs the new timestamped one as in either situation we may
encounter a pre-existing (large) combat log that we need to start
reading from. in both situations, we need to not go directly to the end
of the file, but rather from the point in the file that was written to
before the most recent change(s).

fixes #1767